Add the option to add a secondary subnet to VPC construct
A secondary subnet is mandatory for Application Load Balancers. Availibility
zones for subnets are now set explictly as the secondary subnet needs to have
an availability zone different from the primary one.
The code is also refactored so that the same route table is shared across both
public subnets.

TN: V601-018