Add a way to configure multiple HTTP API stages
Currently HttpApi has a declare_stage function that can be used
to add extra stages, but it doesn't allow to completely configure
new stages with ease, i.e. if you use a domain name and zone. It
is also lacking a way to add lambda permissions as it's done
elsewhere. Here we introduce a new parameter in the constructor
to list and configure all the stages so they are all generated
the same way without having to call an extra function later