Fix flask wrapper for HTTP events
This commit fixes the regression "KeyError: 'identity'" that occurs when
the wrapper must handle an event from an HTTP API. The regression was
introduced on a previous commit, when support for REST API was added.

To avoid such problems in the future this commit also adds a unit test
for HTTP API events.