Add read/write_trusted_accounts parameters
The parameter trusted_accounts of BucketWithRoles only allows
to pass a list of accounts to be trusted both for read access and
write access. However, it may sometimes be desirable to only
trust an account for read access. Here we add two parameters that
can be used to trust additional accounts only for read access or
write access