Add a new version of the VPC construct
This new version allows deploying private subnets in an arbitrary number
of availability zones. Outbound traffic from each private subnet is
routed to a NAT gateway in a corresponding public subnet in the same
availability zone.
Having subnets in at least two different availability zones is required
by Elastic BeanStalk. It is also recommended for any AutoScaling group
to reduce the risk of instance shortages and be more resilient in case
of AZ outages.

ref #2