Add a DeadLetter Config to AWS lambda functions
Some AWS services invoke lambda asynchronously. In this case, the
services invoking lambda only ensure the delivery of the event to the
function. The runtime and timeout errors cannot be handled by the
service.

A deadletter config is added to the `e3.aws.troposphere.awslambda`
attributes to handle event error in case of failure. It can be an
SQS queue or SNS topic.
These settings require to add the statement `sqs:SendMessage`
or `sns:Publish` to the lambda policy according to the deadletter
configuration.

ref it/aws-costs#8
ref it/org/operation_support/iaas/projects#111