Add back and improve build_wheel.py
The call to build_wheel.py is added back after it was removed in
it/e3-aws!27 because of an issue with pushing the new tag.

Now the tag will not be created automatically. Instead, the script
will check that VERSION matches the major/minor version in the
most recent tag that we will have to manually create.

Ref it/org/operation_support/iaas/issues#214