New versioning system for e3 packages
The number of commits since the last tagged version will be
appended to the version of the package as the patch version
by build_wheel.py instead of the date.

The minor version of each package will be bumped in their
respective MRs to avoid conflicts with the old versioning
system.

Ref it/org/operation_support/iaas/issues#214

See merge request it/cookiecutter-e3!60 of the template