Add platforms for AWS Lambda Runtime configurations
As described in the Github [issue](https://github.com/pypa/pip/issues/10760),
`pip` fails to find a manylinux wheel when `--platform` is defined and
there is no exact match.

Add old platforms for the `pip` command when creating the data directory
for the AWS Lambda function.

This should resolved the dependency conflict observed in this
[job](https://gitlab.adacore-it.com/it/iac/-/jobs/864195).