Restore COVERAGE_CORE management in tox.ini
Some of this management was removed at the same time as we removed Gitlab
CI testing for some versions of Python.

However, the Github CI is retained for testing older versions of Python
and needs this management.

In order to have a functional Github CI, this commit restores the
management of this variable, removed by it/cookiecutter-e3!76.

Part of eng/shared/anod#553

See merge request it/cookiecutter-e3!78 of the template