Avoid exception from not modified status code response
Prevent StopIteration exception when Flask app does not return
any response body.

This behavior can be found with the CSS static files.
The static view is automatically added by Flask and returns
a 304 status code in case the files did not change. According
to the documentation, the 304 status code (Not Modified) does
not return any body.