Add an AWS CDK Asset equivalent
The goal is to avoid redeploying resources like Lambdas when their
sources do not change.

Unlike what exists with `s3_data_key`, there will now be
`s3_assets_key`, independent of the timestamp, where assets
will be uploaded. This will allow to check if an asset was
previously deployed and to avoid redeploying it unnecessarily

Ref it/org/operation_support/iaas/projects#11