Add a diff option to the show command
It allows to output the data difference between deployed assets
and the new ones (only PyFunctions use assets currently).

It can also output the difference between the deployed template
body and the new one.

As assets can generate quite long outputs, an option `--assets`
is added to limit the output when not set

Ref it/org/operation_support/iaas/projects#119