Make Bucket.resources not mutate the notifications targets
Each time `Bucket.resources` is called, `notification_setup` is
also called and mutates the Topics and Queues stored in
`topic_configurations` and `queue_configurations`. This can lead
to errors like "Unique Sid is required for TopicPolicy statements".

It also fixes the `TopicPolicy` required to receive S3 notifications
being not added to the template when using `S3WebsiteDistribution`.

As the topic policy was added to the topic only during
`Bucket.resources`, it resulted in `S3WebsiteDistribution.add_cache_invalidation`
returning the resources of the topic without the topic policy