Handle botocore exceptions in PyFunction.diff
It is possible to have a botocore exception depending on
conditions on the IAM policy when trying to get a function that
doesn't exist yet. For example, a policy with tag-based
restrictions. In such cases it's better to catch such exceptions
as it's not possible to know if that's a misconfigured policy
or a function not deployed yet