Move python-on-whales to optional dependencies
Adding python-on-whales to e3-aws dependencies caused e3-distrib to fail. e3-distrib packages with python-on-whales need to be uploaded on cathod to solve this.

We may not want to add python-on-whales on cathod because it is just used for GNAT IQ.

Moving python-on-whales to optional dependencies resolves the e3-distrib build issue.

ref: https://gitlab.adacore-it.com/it/org/software-supply-chain/production-pipeline/issues/-/issues/192